[#1776] Move Membership Formset base view out
It's now in it's own file for base views.
