+++ amy/fiscal/views.py
@@ -12,9 +12,7 @@ from django.db.models import (
 )
 from django.db.models.functions import Now
 from django.forms import modelformset_factory
-from django.shortcuts import get_object_or_404
 from django.urls import reverse, reverse_lazy
-from django.views.generic import FormView
 
 from fiscal.filters import (
     OrganizationFilter,
@@ -30,6 +28,9 @@ from fiscal.forms import (
     SponsorshipForm,
 )
 from fiscal.models import MembershipTask
+from fiscal.base_views import (
+    MembershipFormsetView,
+)
 from workshops.base_views import (
     AMYCreateView,
     AMYUpdateView,
@@ -248,50 +249,6 @@ class MembershipDelete(OnlyForAdminsMixin, PermissionRequiredMixin, AMYDeleteVie
 
     def get_success_url(self):
         return reverse("all_memberships")
-
-
-class MembershipFormsetView(FormView):
-    template_name = "fiscal/membership_formset.html"
-
-    def dispatch(self, request, *args, **kwargs):
-        self.membership = get_object_or_404(Membership, pk=self.kwargs["membership_id"])
-        return super().dispatch(request, *args, **kwargs)
-
-    def get_formset_kwargs(self):
-        return {
-            "extra": 0,
-            "can_delete": True,
-        }
-
-    def get_form_class(self):
-        return self.get_formset(**self.get_formset_kwargs())
-
-    def get_form_kwargs(self):
-        kwargs = super().get_form_kwargs()
-        kwargs["queryset"] = self.get_formset_queryset(self.membership)
-        return kwargs
-
-    def get_context_data(self, **kwargs):
-        kwargs["membership"] = self.membership
-        kwargs["formset"] = self.get_form()
-        return super().get_context_data(**kwargs)
-
-    def form_valid(self, form):
-        instances = form.save(commit=False)
-
-        # assign membership to any new/changed instance
-        for instance in instances:
-            instance.membership = self.membership
-            instance.save()
-
-        # remove deleted objects
-        for instance in form.deleted_objects:
-            instance.delete()
-
-        return super().form_valid(form)
-
-    def get_success_url(self):
-        return self.membership.get_absolute_url()
 
 
 class MembershipMembers(OnlyForAdminsMixin, MembershipFormsetView):
